move csv component to utils directory
